% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combin_outputs.R
\name{combin.mo.ind}
\alias{combin.mo.ind}
\title{Combining outputs from separate runs of function 'morpho.ind'}
\usage{
combin.mo.ind(out1 = out1, out2 = out2)
}
\description{
This is to be used ONLY if both runs had the same reference sample. Generally it wwill be used when mixed data is used and one run of morpho.ind will have parametric bootstraps while the other has non-parametric bootstrap.
}
