% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combin_outputs.R
\name{combin_mo_ind}
\alias{combin_mo_ind}
\title{Combining outputs from separate runs of function 'morpho_indices'}
\usage{
combin_mo_ind(out1 = out1, out2 = out2)
}
\description{
This is to be used ONLY if both runs had the same reference sample. Generally it wwill be used when mixed data is used and one run of morpho_indices will have parametric bootstraps while the other has non-parametric bootstrap.
}
