% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_indices.R
\name{plot_mo_ind}
\alias{plot_mo_ind}
\title{Plotting function for 'morpho_indices' outputs}
\usage{
plot_mo_ind(
  dat = dat,
  name = name,
  ref = 1,
  samp = name[-ref],
  var.names = colnames(dat$LSI.df),
  which.var = 1:length(var.names),
  index = "LSI",
  plot.ref = T,
  plot.index = T,
  bootmean = F,
  bootmedian = F,
  plot.mISD = T,
  CI = F,
  col.ref = "black",
  col = rainbow(length(samp)),
  alpha.lev = 0.2,
  lty.index = 1,
  lty.bootmean = 2,
  lty.bootmed = 3,
  lgnd = T,
  lgnd.pos = "topleft",
  cex.lgnd = 0.5,
  xlab = "Descriptors",
  ylab = index,
  tick.labels = colnames(dfs)
)
}
\arguments{
\item{dat}{Data is the output from function morpho.indices().}

\item{name}{Character vector of names for ALL samples INCLUDING reference.}

\item{ref}{Index number (or name as character string) of reference sample. Default = 1, so the first sample is assumed to be the reference.}

\item{samp}{Numeric (or character) vector containing the index numbers (or names) of samples to be plotted. Defaults to all except reference sample.}

\item{var.names}{Character vector containing the names of ALL morphological descriptors. By default the column names of data frame from the "dat" argument.}

\item{which.var}{Numeric (or character) vector specifying the morphological descriptors to be plotted. By default all of them.}

\item{index}{Which morphological index will be plotted: "LSI" (default), "VSI" or "VSI2".}

\item{plot.ref}{Plot reference sample? Default = TRUE}

\item{plot.index}{Plot original values of LSI, VSI or VSI2. Default = TRUE.}

\item{bootmean}{Plot average bootstraped value of LSI, VSI or VSI2. Default = FALSE.}

\item{bootmedian}{Plot median bootstraped value of LSI, VSI or VSI2. Default = FALSE.}

\item{plot.mISD}{Should mISD values be plotted?}

\item{CI}{Plot 95 confidence intervals? Default = FALSE.}

\item{col.ref}{Color for reference sample. Default is "black".}

\item{col}{Vector of colors for samples to be plotted. Length must be equal to length(samp).}

\item{alpha.lev}{Opacity level (0 = transparent, 1 = opaque). Requires packages 'scales'.}

\item{lty.index, }{lty.bootmean, lty.bootmed Line types for index, bootstrap mean and bootstrap median, respectively.}

\item{lgnd}{Plot legend? Default=TRUE.}

\item{lgnd.pos}{Position of legend, can be specified as for function legend(). Default = 'topleft'.}

\item{cex.lgnd}{Size of legend. Default = 0.5.}

\item{xlab, }{ylab As for standard 'plot'.}

\item{tick.labels}{Labels for tickmarks.}
}
\description{
Plotting function for 'morpho_indices' outputs
}
